Add simple auto boxing for ints in expression evaluator
