+++ core/src/main/scala/ch/epfl/scala/debugadapter/internal/evaluator/package.scala
@@ -3,7 +3,7 @@ package ch.epfl.scala.debugadapter.internal
 import com.sun.jdi._
 
 import scala.collection.JavaConverters._
-import scala.util.Try
+import scala.util.{Failure, Success, Try}
 
 package object evaluator {
   private[evaluator] def method(name: String, referenceType: ReferenceType) =
@@ -13,7 +13,12 @@ package object evaluator {
     Try(referenceType.methodsByName(name, signature).asScala.headOption).toOption.flatten
 
   private[evaluator] def invokeMethod(objRef: ObjectReference, method: Method, args: List[Value], thread: ThreadReference) =
-    Try(objRef.invokeMethod(thread, method, args.asJava, ObjectReference.INVOKE_SINGLE_THREADED)).toOption
+    Try(objRef.invokeMethod(thread, method, args.asJava, ObjectReference.INVOKE_SINGLE_THREADED)) match {
+      case Failure(error: InvocationException) =>
+        println(error.exception())
+        None
+      case Success(result) => Some(result)
+    }
 
   private[evaluator] def classLoader(objRef: ObjectReference): Option[ClassLoaderReference] =
     Try(objRef.referenceType().classLoader()).toOption
