+++ core/src/test/resources/EvaluateTest.scala
@@ -1,7 +1,7 @@
 object EvaluateTest {
   def main(args: Array[String]): Unit = {
-    val a: Integer = 1
-    val b: Integer = 2
+    val a: Int = 1
+    val b: Int = 2
     val c: String = "1 + 2 = "
     println(c + (a + b))
   }
